Fixed ClassLoading Problems with OSGI-Bundles in Eclipse (fixes #56)
The Linked ClassLoader, used in the ClassProxyFactory, calls now two
different ClassLoaders to load the Class to Mock. First, the ClassLoader
of the mocked Class is used. If this ClassLoader cannot load (indicated
by a ClassNotFoundException the given Class, the Classloader of EasyMock
is used.

A automated test for this change will be added after some additional
testing.